requirements.txt
MetaTrader5
anthropic
openai
pandas
requests
python-dotenv
tzdata
backports.zoneinfo; python_version < "3.9"
